requirements.txt
asgiref==3.8.1
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
